フロントのページ修正
