+++ workspace/src/components/artist-view.tsx
