+++ workspace/src/components/name-id.tsx
@@ -23,5 +23,4 @@ const NameId:React.FC<NameIdProps> = ({name,id}) => {
     </div>
   )
 }
-
 export default NameId
