+++ workspace/src/components/posts-view.tsx
