+++ workspace/src/components/track-view.tsx
